Setup numbers right

+++ database/migrations/2017_06_10_112246_create_events_table.php
@@ -16,7 +16,7 @@ class CreateEventsTable extends Migration
         Schema::create('events', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name');
-            $table->float('price', 10, 2);
+            $table->decimal('price', 10, 2);
             $table->date('event_date');
             $table->timestamps();
         });
