+++ database/migrations/2017_06_10_114244_create_products_table.php
@@ -16,7 +16,7 @@ class CreateProductsTable extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name');
-            $table->float('price', 10, 2);
+            $table->decimal('price', 10, 2);
             $table->timestamps();
         });
     }
